+++ src/api/one/scheme/index.ts
@@ -25,7 +25,7 @@ import { GetPostById } from './square/get-post-by-id'
 import { Action as GetPostLikesRaw } from './square/like/get-post-likes.action/interface'
 import { Action as GetPostsOfBoardRaw } from './square/get-posts-of-board.action/interface'
 import { Action as GetRecentPostsOfBoardRaw } from './square/get-recent-posts-of-board.action/interface'
-import { Action as GetTipTopicRaw } from './honey_tip/get-topics.action/interface'
+import { Action as GetTipTopicRaw } from './tip/get-topics.action/interface'
 import { Action as GetUserIdRaw } from './auth/get-user-id.action/interface'
 import { Action as IsPostLikedRaw } from './square/like/is-post-liked.action/interface'
 import { Action as LikePostRaw } from './square/like/like-post.action/interface'
